Separated into DateFunctionReturnTypeHelper
