+++ conf/config.neon
@@ -1273,6 +1273,9 @@ services:
 			- phpstan.broker.dynamicFunctionReturnTypeExtension
 
 	-
+		class: PHPStan\Type\Php\DateFunctionReturnTypeHelper
+
+	-
 		class: PHPStan\Type\Php\DateFormatFunctionReturnTypeExtension
 		tags:
 			- phpstan.broker.dynamicFunctionReturnTypeExtension
